move abstractions and presenters to core, make empty NextStep view and presenter
