+++ app/src/main/java/org/stepic/droid/core/CourseDetailModule.java
@@ -6,8 +6,8 @@ import com.squareup.otto.Bus;
 import org.stepic.droid.concurrency.IMainHandler;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.store.operations.DatabaseFacade;
-import org.stepic.droid.ui.presenters.course_finder.CourseFinderPresenter;
-import org.stepic.droid.ui.presenters.course_joiner.CourseJoinerPresenter;
+import org.stepic.droid.core.presenters.CourseFinderPresenter;
+import org.stepic.droid.core.presenters.CourseJoinerPresenter;
 import org.stepic.droid.web.IApi;
 
 import java.util.concurrent.ThreadPoolExecutor;
