+++ app/src/main/java/org/stepic/droid/core/SectionModule.java
@@ -11,8 +11,8 @@ import org.stepic.droid.configuration.IConfig;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.preferences.UserPreferences;
 import org.stepic.droid.store.operations.DatabaseFacade;
-import org.stepic.droid.ui.presenters.course_finder.CourseFinderPresenter;
-import org.stepic.droid.ui.presenters.course_joiner.CourseJoinerPresenter;
+import org.stepic.droid.core.presenters.CourseFinderPresenter;
+import org.stepic.droid.core.presenters.CourseJoinerPresenter;
 import org.stepic.droid.web.IApi;
 
 import java.util.concurrent.ThreadPoolExecutor;
