+++ app/src/main/java/org/stepic/droid/core/StepComponent.kt
@@ -1,0 +1,10 @@
+package org.stepic.droid.core
+
+import dagger.Subcomponent
+import org.stepic.droid.base.StepBaseFragment
+
+@PerFragment
+@Subcomponent(modules = arrayOf(StepModule::class))
+interface StepComponent {
+    fun inject(stepFragment: StepBaseFragment)
+}
