+++ app/src/main/java/org/stepic/droid/core/StepModule.java
@@ -1,0 +1,15 @@
+package org.stepic.droid.core;
+
+import org.stepic.droid.core.presenters.NextStepPresenter;
+
+import dagger.Module;
+import dagger.Provides;
+
+@Module
+public class StepModule {
+    @Provides
+    public NextStepPresenter provideNextStepPresenter() {
+        return new NextStepPresenter();
+    }
+
+}
