+++ app/src/main/java/org/stepic/droid/core/presenters/NextStepPresenter.kt
@@ -1,0 +1,14 @@
+package org.stepic.droid.core.presenters
+
+import org.stepic.droid.core.presenters.contracts.NextStepView
+import org.stepic.droid.model.Lesson
+
+class NextStepPresenter : PresenterBase<NextStepView>() {
+
+    /**
+     * Last step in lesson can be shown differently
+     */
+    fun checkStepForLast(stepId: Long, lesson: Lesson) {
+
+    }
+}
