+++ app/src/main/java/org/stepic/droid/core/presenters/PresenterBase.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.ui.presenters
+package org.stepic.droid.core.presenters
 
 import android.support.annotation.CallSuper
