+++ app/src/main/java/org/stepic/droid/core/presenters/contracts/NextStepView.java
@@ -1,0 +1,4 @@
+package org.stepic.droid.core.presenters.contracts;
+
+public interface NextStepView {
+}
