+++ app/src/main/java/org/stepic/droid/ui/adapters/CertificateAdapter.java
@@ -18,7 +18,7 @@ import org.stepic.droid.R;
 import org.stepic.droid.base.MainApplication;
 import org.stepic.droid.model.CertificateType;
 import org.stepic.droid.model.CertificateViewItem;
-import org.stepic.droid.ui.presenters.certificate.CertificatePresenter;
+import org.stepic.droid.core.presenters.CertificatePresenter;
 
 import java.util.List;
