+++ app/src/main/java/org/stepic/droid/ui/fragments/SectionFragment.java
@@ -46,6 +46,10 @@ import org.stepic.droid.core.CalendarExportableView;
 import org.stepic.droid.core.CalendarPresenter;
 import org.stepic.droid.core.SectionModule;
 import org.stepic.droid.core.ShareHelper;
+import org.stepic.droid.core.presenters.CourseFinderPresenter;
+import org.stepic.droid.core.presenters.contracts.CourseJoinView;
+import org.stepic.droid.core.presenters.CourseJoinerPresenter;
+import org.stepic.droid.core.presenters.contracts.LoadCourseView;
 import org.stepic.droid.events.CalendarChosenEvent;
 import org.stepic.droid.events.courses.CourseCantLoadEvent;
 import org.stepic.droid.events.courses.CourseFoundEvent;
@@ -64,15 +68,11 @@ import org.stepic.droid.model.Course;
 import org.stepic.droid.model.Section;
 import org.stepic.droid.notifications.INotificationManager;
 import org.stepic.droid.notifications.model.Notification;
-import org.stepic.droid.ui.abstraction.CourseJoinView;
-import org.stepic.droid.ui.abstraction.LoadCourseView;
 import org.stepic.droid.ui.adapters.SectionAdapter;
 import org.stepic.droid.ui.dialogs.ChooseCalendarDialog;
 import org.stepic.droid.ui.dialogs.ExplainCalendarPermissionDialog;
 import org.stepic.droid.ui.dialogs.LoadingProgressDialog;
 import org.stepic.droid.ui.dialogs.UnauthorizedDialogFragment;
-import org.stepic.droid.ui.presenters.course_finder.CourseFinderPresenter;
-import org.stepic.droid.ui.presenters.course_joiner.CourseJoinerPresenter;
 import org.stepic.droid.util.AppConstants;
 import org.stepic.droid.util.ColorUtil;
 import org.stepic.droid.util.HtmlHelper;
